Use SourceClient directly in functional CreateGitSourceTest
